Fixed design
Fixed the design and now uses the universal css folder, also expands tabs. UpdateAgent has similar updates to agenda.

Also, the styles.css is identical to the one in main with one new addition this addition is textarea on line 83. This will fix the resizable message box.

Lastly, the scripts.js is just copied and pasted from main so it is identical to the main.

I didnt know if I had to change these files so let me know if I shouldve left them alone.